chore(tests): move utils one level higher
